server/rss-fetcher/requirements.txt
newspaper4k
sqlalchemy
feedparser
requests
lxml[html_clean]